Fix growths not visible in custom cam
